backend/pyproject.toml
[project]
name = "duotrak-backend"
version = "1.1.0"
description = "AI orchestration backend for DuoTrak"
readme = "backend_implentation.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite",
  "alembic",
  "asyncpg",
  "boto3",
  "crewai==1.9.3",
  "fastapi",
  "google-genai==1.64.0",
  "httpx",
  "langchain-google-genai==4.2.1",
  "numpy",
  "pinecone-client",
  "psycopg2-binary",
  "pydantic-settings",
  "PyJWT",
  "python-dotenv",
  "python-multipart",
  "redis",
  "slowapi",
  "sqlalchemy[asyncio]",
  "tenacity",
  "typing-extensions",
  "uvicorn[standard]",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
